Fix initial ammo assignment
Before this patch, player is assigned bullets (1) first through the inspector, (2) then again 15 times
that amount in addition when the game starts, and (3) finally a calculated amount that equalizes the
amount of player's ammo and the total amount of enemies that will spawn in the upcoming round.

As a result of (3), (1) and (2) are completely redundant, since (3) will always overwrite the previous
value. Also, all three together can result in (1) and (2) assigning a value higher than the amount of
total enemies, and instead of (3) happening, player will be awarded points before even started playing
the game.

After the patch, ammo is only assigned before the round starts, and also, a bonus of 55% more ammo is
given to the player before the 1st round to allow them some slack for missed shots.

Signed-off-by: AnimaRain <ioannis_spyr@yahoo.gr>